feat: add registration and authentication functionality
First version of login and registration feature based on JWT tokens.
Add this functionality with the help of dj-rest-auth, rest-framework-simplejwt and allauth libraries.
Both access and refresh tokens are put into httponly cookies. Add custom views responsible for setting and unsetting cookies while logging in and out.
Also add custom serializers based on those from dj-rest-auth in order not to send access and refresh tokens in response body.
Had to change account app label because of the conflict with allauth.account app in installed apps.
Add setting for CamelCase renderer to put no underscore before number (needed to send password1 and password2 not password_1 and password_2 in registration)